+++ CHANGELOG.md
@@ -104,7 +104,12 @@ best friend, lajbel, can put the correct version name here
   });
   ```
 
+- Added `skew` to text formatting, so now italics is possible - @dragoncoder047
+
 ### Changed
+
+- (**!**) Renamed `onShow()` to `onTabShow()` and `onHide()` to `onTabHide()` -
+  @lajbel
 
 - In addition to being the `scene()` function, now `scene` is also a scope for
   scene event handlers - @lajbel